feat: refactored to move geo and lat long to internal
